refactor(acp): extract transcript row views
