+++ Alas/Sources/ACP/UI/ACPTranscriptQueuePolicy.swift
@@ -1,0 +1,27 @@
+import Foundation
+
+/// Which queued prompts get a bubble in the transcript's synthetic tail, and
+/// which drags between them are legal.
+enum ACPTranscriptQueuePolicy {
+    nonisolated static func shouldRenderQueueBubble(status: QueuedPrompt.Status) -> Bool {
+        switch status {
+        case .pending:
+            return true
+        case .sending:
+            return false
+        }
+    }
+
+    nonisolated static func queueHeaderCount(statuses: [QueuedPrompt.Status]) -> Int {
+        statuses.reduce(0) { count, status in
+            count + (shouldRenderQueueBubble(status: status) ? 1 : 0)
+        }
+    }
+
+    nonisolated static func canDropQueuedItem(
+        sourceStatus: QueuedPrompt.Status?,
+        targetStatus: QueuedPrompt.Status
+    ) -> Bool {
+        sourceStatus == .pending && targetStatus == .pending
+    }
+}
